Change min_participants type

+++ db/schema.rb
@@ -23,7 +23,7 @@ ActiveRecord::Schema.define(version: 20140427123755) do
     t.string   "picture"
     t.boolean  "approved",         default: false
     t.integer  "lecturer_id"
-    t.decimal  "min_participants", default: 0.0
+    t.integer  "min_participants", default: 0
     t.integer  "location_id"
   end
